// Atspm/WebUI/src/api/config/aTSPMConfigurationApi.schemas.ts
export interface Approach {
  /** @nullable */
  description?: string | null
  /** @nullable */
  mph?: number | null
  protectedPhaseNumber?: number
  isProtectedPhaseOverlap?: boolean
  /** @nullable */
  permissivePhaseNumber?: number | null
  isPermissivePhaseOverlap?: boolean
  /** @nullable */
  pedestrianPhaseNumber?: number | null
  isPedestrianPhaseOverlap?: boolean
  /** @nullable */
  pedestrianDetectors?: string | null
  locationId?: number
  location?: Location
  directionTypeId?: DirectionTypes
  directionType?: DirectionType
  /** @nullable */
  detectors?: Detector[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
}
// ...
export interface Area {
  /** @nullable */
  name?: string | null
  /** @nullable */
  userAreas?: UserArea[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}

export interface ArrivalOnRedOptions {
  /** @nullable */
  locationIdentifier?: string | null
  start?: string
  end?: string
  binSize?: number
  getPermissivePhase?: boolean
}

export interface AtspmOptionsBase {
  [key: string]: unknown
}

export type DetectionHardwareTypes =
  (typeof DetectionHardwareTypes)[keyof typeof DetectionHardwareTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DetectionHardwareTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_7: 7,
} as const

export interface DetectionType {
  /** @nullable */
  description?: string | null
  /** @nullable */
  abbreviation?: string | null
  displayOrder?: number
  /** @nullable */
  measureTypes?: MeasureType[] | null
  id?: DetectionTypes
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  detectors?: Detector[] | null
}

export interface DetectionTypeDto {
  id?: DetectionTypes
  /** @nullable */
  description?: string | null
  /** @nullable */
  abbreviation?: string | null
  displayOrder?: number
  /** @nullable */
  measureTypes?: MeasureTypeDto[] | null
}

export interface DetectionTypeGroup {
  /** @nullable */
  id?: string | null
  count?: number
}

export type DetectionTypes =
  (typeof DetectionTypes)[keyof typeof DetectionTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DetectionTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_7: 7,
  NUMBER_8: 8,
  NUMBER_9: 9,
  NUMBER_10: 10,
  NUMBER_11: 11,
} as const

export interface Detector {
  /** @nullable */
  dectectorIdentifier?: string | null
  detectorChannel?: number
  /** @nullable */
  distanceFromStopBar?: number | null
  /** @nullable */
  minSpeedFilter?: number | null
  dateAdded?: string
  /** @nullable */
  dateDisabled?: string | null
  /** @nullable */
  laneNumber?: number | null
  movementType?: MovementTypes
  laneType?: LaneTypes
  detectionHardware?: DetectionHardwareTypes
  /** @nullable */
  decisionPoint?: number | null
  /** @nullable */
  movementDelay?: number | null
  latencyCorrection?: number
  approachId?: number
  /** @nullable */
  detectorComments?: DetectorComment[] | null
  /** @nullable */
  detectionTypes?: DetectionType[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  approach?: Approach
}

export interface DetectorComment {
  timeStamp?: string
  /** @nullable */
  comment?: string | null
  detectorId?: number
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  detector?: Detector
}
// ...
export interface Device {
  /** @nullable */
  deviceIdentifier?: string | null
  loggingEnabled?: boolean
  /** @nullable */
  ipaddress?: string | null
  deviceStatus?: DeviceStatus
  deviceType?: DeviceTypes
  /** @nullable */
  notes?: string | null
  locationId?: number
  /** @nullable */
  deviceConfigurationId?: number | null
  deviceConfiguration?: DeviceConfiguration
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  location?: Location
}

export interface DeviceConfiguration {
  /** @nullable */
  description?: string | null
  /** @nullable */
  notes?: string | null
  protocol?: TransportProtocols
  port?: number
  /** @nullable */
  path?: string | null
  /** @nullable */
  query?: string[] | null
  connectionTimeout?: number
  operationTimeout?: number
  loggingOffset?: number
  /** @nullable */
  decoders?: string[] | null
  /** @nullable */
  userName?: string | null
  /** @nullable */
  password?: string | null
  /** @nullable */
  productId?: number | null
  product?: Product
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  devices?: Device[] | null
}

export interface DeviceGroup {
  /** @nullable */
  manufacturer?: string | null
  /** @nullable */
  model?: string | null
  /** @nullable */
  firmware?: string | null
  count?: number
}

export type DeviceStatus = (typeof DeviceStatus)[keyof typeof DeviceStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeviceStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const

export type DeviceTypes = (typeof DeviceTypes)[keyof typeof DeviceTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeviceTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
} as const

export interface DirectionType {
  /** @nullable */
  description?: string | null
  /** @nullable */
  abbreviation?: string | null
  displayOrder?: number
  /** @nullable */
  primaryDirections?: RouteLocation[] | null
  id?: DirectionTypes
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  opposingDirections?: RouteLocation[] | null
  /** @nullable */
  approaches?: Approach[] | null
}

export type DirectionTypes =
  (typeof DirectionTypes)[keyof typeof DirectionTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DirectionTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_7: 7,
  NUMBER_8: 8,
} as const
// ...
export interface Jurisdiction {
  /** @nullable */
  name?: string | null
  /** @nullable */
  mpo?: string | null
  /** @nullable */
  countyParish?: string | null
  /** @nullable */
  otherPartners?: string | null
  /** @nullable */
  userJurisdictions?: UserJurisdiction[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}

export type LaneTypes = (typeof LaneTypes)[keyof typeof LaneTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LaneTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_7: 7,
} as const
// ...
export interface Location {
  latitude?: number
  longitude?: number
  /** @nullable */
  primaryName?: string | null
  /** @nullable */
  secondaryName?: string | null
  chartEnabled?: boolean
  versionAction?: LocationVersionActions
  /** @nullable */
  note?: string | null
  start?: string
  pedsAre1to1?: boolean
  /** @nullable */
  locationIdentifier?: string | null
  /** @nullable */
  jurisdictionId?: number | null
  jurisdiction?: Jurisdiction
  locationTypeId?: number
  locationType?: LocationType
  /** @nullable */
  regionId?: number | null
  region?: Region
  /** @nullable */
  areas?: Area[] | null
  /** @nullable */
  devices?: Device[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  approaches?: Approach[] | null
}
// ...
export interface LocationType {
  /** @nullable */
  name?: string | null
  /** @nullable */
  icon?: string | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}

export type LocationVersionActions =
  (typeof LocationVersionActions)[keyof typeof LocationVersionActions]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LocationVersionActions = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_10: 10,
} as const
// ...
export interface MeasureType {
  /** @nullable */
  name?: string | null
  /** @nullable */
  abbreviation?: string | null
  showOnWebsite?: boolean
  showOnAggregationSite?: boolean
  displayOrder?: number
  /** @nullable */
  measureComments?: MeasureComment[] | null
  /** @nullable */
  measureOptions?: MeasureOption[] | null
  /** @nullable */
  measureOptionPresets?: MeasureOptionPreset[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  detectionTypes?: DetectionType[] | null
}
// ...
export type MovementTypes = (typeof MovementTypes)[keyof typeof MovementTypes]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MovementTypes = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6,
  NUMBER_7: 7,
} as const
// ...
export interface Region {
  /** @nullable */
  description?: string | null
  /** @nullable */
  userRegions?: UserRegion[] | null
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  /** @nullable */
  locations?: Location[] | null
}
// ...
export interface RouteLocation {
  order?: number
  primaryPhase?: number
  opposingPhase?: number
  primaryDirectionId?: DirectionTypes
  opposingDirectionId?: DirectionTypes
  isPrimaryOverlap?: boolean
  isOpposingOverlap?: boolean
  /** @nullable */
  previousLocationDistanceId?: number | null
  previousLocationDistance?: RouteDistance
  /** @nullable */
  nextLocationDistanceId?: number | null
  /** @nullable */
  locationIdentifier?: string | null
  routeId?: number
  route?: Route
  id?: number
  /** @nullable */
  created?: string | null
  /** @nullable */
  modified?: string | null
  /** @nullable */
  createdBy?: string | null
  /** @nullable */
  modifiedBy?: string | null
  primaryDirection?: DirectionType
  opposingDirection?: DirectionType
  nextLocationDistance?: RouteDistance
}
// ...
export interface UserArea {
  /** @nullable */
  userId?: string | null
  areaId?: number
  area?: Area
}

export interface UserJurisdiction {
  /** @nullable */
  userId?: string | null
  jurisdictionId?: number
  jurisdiction?: Jurisdiction
}

export interface UserRegion {
  /** @nullable */
  userId?: string | null
  regionId?: number
  region?: Region
}